pyproject.toml
[project]
name = "flatwhite"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pytrends>=4.9.2",
    "yfinance>=0.2.36",
    "feedparser>=6.0.11",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.3",
    "numpy>=1.26.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    "google-genai>=1.0.0",
    "anthropic>=0.40.0",
    "playwright>=1.40.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "youtube-transcript-api>=0.6.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.0.0",
    "openpyxl>=3.1.0",
]

[project.scripts]
flatwhite = "flatwhite.cli:main"

[tool.setuptools.packages.find]
include = ["flatwhite*"]
